Refactored admin dashboard components and routes
