+++ src/main.jsx
@@ -4,12 +4,14 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import './index.css';
 import HomePage from './components/HomePage';
 import AdminLogin from './components/adminportal/AdminLogin';
-import AdminDashboard from './components/adminportal/AdminDashboard';
+// import AdminDashboard from './components/adminportal/AdminDashboard';
 import IITRoleLogin from './components/logins/iitlogins/IITRoleLogin.jsx';
 import UniversityLoginDashboard from './components/logins/UniversityLoginDashboard.jsx';
 import IITStudentLogin from './components/logins/iitlogins/IITStudentLogin.jsx';
 import TimeTableUpdate from './components/iitadmin/TimeTableUpdate.jsx';
 import InquiryForm from './components/InquiryForm.jsx';
+import AdminPortalAdminDashboard from './components/adminportal/AdminPortalAdminDashboard.jsx';
+import AdminDashboard from './components/dashboard/iit/AdminDashboard.jsx';
 
 
 
@@ -19,12 +21,13 @@ createRoot(document.getElementById('root')).render(
       <Routes>
         <Route path="/" element={<HomePage />} />
         <Route path="/admin-login" element={<AdminLogin />} />
-        <Route path="/admin-dashboard" element={<AdminDashboard />} />
+        <Route path="/admin-dashboard" element={<AdminPortalAdminDashboard />} />
         <Route path="/uni-login" element={<UniversityLoginDashboard />} />
         <Route path="/iit-role-login" element={<IITRoleLogin />} />
         <Route path="/iit-student-login" element={<IITStudentLogin />} />
         <Route path="/timetable-update" element={<TimeTableUpdate />} />
         <Route path="/inquiry" element={<InquiryForm />} />
+        <Route path="/a-dashboard" element={<AdminDashboard/>} />
 
 
       </Routes>
